Add files via upload
A stored procedure is required to copy data from the Job Operation table into a new table and is run in a MSSQL job at the beginning of each shift (6am and 6pm). Because of the way the database was set up there is no historical information on what was originally planned. Each job is scheduled to start and finish at specific times. Once the operation has started the original planned start time is overwritten.  By copying data into a new table before the operation is scheduled to start, we can compare it later with actual results.