Create IDC Data Transformation
The purpose of these stored procedures are to minimise the the manual cleaning of IDC records. An SSRS report was set up to run the second stored procedure and issue the results to staff in an email.

The issue with SSRS is that it will only read the data from the first table, whereas we need the transformed data from the second. To get around this I changed the temp tables to global temp tables and left them open at the end of the session. A SQL server agent job was set up to run the first stored procedure at 6am. SSRS is then set up to run the second procedure which selects from the target results data, then closes both tables.

On test this works in my current environment, however the SSRS report subscription fails because global temp tables automatically close when a session has ended....in this case the SQL server job.  So when the second procedure is run, it fails because the temp tables from the first procedure don't exist.